requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.12.1
djoser==2.0.5
idna==2.10
Markdown==3.3.3
python-decouple==3.3
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
